Revert "Add support for New York city's Taxi and Limousine Commission yellow taxi trip records (#151)"
We decided not to use Apache Parquet data. Instead of it, we can use it by Red Datasets Parquet.
The reason is why we don't want to add red-parquet dependency to red-datasets for easy to install.
red-parquet is an extension library. So it's difficult to install rather than a pure Ruby library.