Revert "Add support for New York city's Taxi and Limousine Commission yellow taxi trip records" (#154)
## What I did
- Reverted https://github.com/red-data-tools/red-datasets/pull/151

### Reasons
We decided not to use Apache Parquet data in Red Datasets because we
don't want to add red-parquet dependency to red-datasets for easy to
install.
In details, red-parquet is an extension library. So it's difficult to
install rather than a pure Ruby library.
Instead of it, we created [Red Datasets
Parquet](https://github.com/red-data-tools/red-datasets-parquet) to use
Apache Parquet data.