Prepare to introduce rbs and add a configuration file (#238)
### What I did in this PR
To introduce rbs to this repository, I added the necessary gems like
`steep` and `typeprof` to gemspec file.
I use these two gems to get support for the RBS file's type checking and
type generation.
I also added a Steepfile, which includes how to check the RBS files. The
configuration of Steepfile is minimal, such as which directory to put
RBS files in and not to raise an error with type checking at first.

### What I didn't do in this PR
This is a PR for preparation to introduce RBS to this repository, I
added only configuration files.
The actual RBS file will be added in the next PR with a new dataset.